refactor: 링크 필요없는 로직 제거, 유틸 로직 분리
